+++ PiLights/DeviceServer/Model/Thermometer.swift
@@ -10,7 +10,7 @@ import Foundation
 import SQLite
 
 class Thermometer: Device {
-    init(id: Int, name: String, value: Int, roomId: Int) {
+    init(id: Int, name: String, value: Double, roomId: Int) {
         super.init(id: id, name: name, type: .thermometer, value: value, roomId: roomId)
     }
 
@@ -19,7 +19,7 @@ class Thermometer: Device {
         self.init(
             id: try values.decode(Int.self, forKey: .id),
             name: try values.decode(String.self, forKey: .name),
-            value: try values.decode(Int.self, forKey: .value),
+            value: try values.decode(Double.self, forKey: .value),
             roomId: try values.decode(Int.self, forKey: .roomId)
         )
     }
@@ -28,7 +28,7 @@ class Thermometer: Device {
         self.init(
             id: Int(row[0] as! Int64),
             name: row[1] as! String,
-            value: Int(row[4] as! Int64),
+            value: row[4] as! Double,
             roomId: Int(row[3] as! Int64)
         )
     }
