+++ PiLights/DeviceServer/Networking/DeviceStatusPacket.swift
@@ -10,9 +10,9 @@ import Foundation
 
 class DeviceStatusPacket: Codable {
     public let deviceId: Int
-    public let deviceValue: Int
+    public let deviceValue: Double
 
-    public init(deviceId: Int, deviceValue: Int) {
+    public init(deviceId: Int, deviceValue: Double) {
         self.deviceId = deviceId
         self.deviceValue = deviceValue
     }
